+++ pkg/generator/generator.go
@@ -4,8 +4,6 @@ import (
 	"bytes"
 	"fmt"
 	"github.com/Ygg-Drasill/Sleipnir/pkg/ast"
-	"github.com/Ygg-Drasill/Sleipnir/pkg/gocc/token"
-	"log/slog"
 	"reflect"
 )
 
@@ -61,70 +59,6 @@ func (g *Generator) genProgram(node *ast.Program) string {
 	return ""
 }
 
-func (g *Generator) genNode(node *ast.Node) string {
-	g.currentNode = node
-	g.write("(func $%s\n", node.Id)
-	g.genNodeLocals(&node.ProcStatements)
-	g.genNodeGlobals(&node.OutDeclarations)
-
-	inputs := make([]ast.Connection, 0)
-	for _, conn := range g.syntaxTree.Connections {
-		if conn.InJunction.NodeId != node.Id {
-			continue
-		}
-		inputs = append(inputs, conn)
-	}
-
-	connectionsMemo := make(map[string]bool)
-	for _, conn := range inputs {
-		if connectionsMemo[conn.InJunction.NodeId] {
-			break
-		}
-		g.write("global.get $%s_processed\n", conn.OutJunction.NodeId)
-		g.write("(if (then nop) (else return))\n")
-		connectionsMemo[conn.InJunction.NodeId] = true
-	}
-
-	for _, stmt := range node.ProcStatements {
-		g.gen(&stmt)
-	}
-
-	clear(connectionsMemo)
-	connectionsMemo = make(map[string]bool)
-	for _, conn := range g.syntaxTree.Connections {
-		if conn.OutJunction.NodeId == node.Id && !connectionsMemo[conn.InJunction.NodeId] {
-			g.write("call $%s\n", conn.InJunction.NodeId)
-			connectionsMemo[conn.InJunction.NodeId] = true
-		}
-	}
-	g.write(")\n")
-	return ""
-}
-
-func (g *Generator) genNodeLocals(statements *ast.StatementList) string {
-	for _, statement := range *statements {
-		dec, ok := statement.(ast.Declaration)
-		if !ok {
-			continue
-		}
-		g.write("(local $%s i32)\n", dec.AssigneeId)
-	}
-	return ""
-}
-
-func (g *Generator) genNodeGlobals(decList *ast.DeclarationList) string {
-	for _, assignment := range *decList {
-		value := 0
-		if v, ok := assignment.Expression.(int64); assignment.Expression != nil && ok {
-			value = int(v)
-		}
-
-		g.write("(global.set $%s_%s (i32.const %d))\n", g.currentNode.Id, assignment.AssigneeId, value)
-	}
-
-	return ""
-}
-
 func (g *Generator) genStmt(node *ast.Statement) string {
 	if assStmt, ok := (*node).(ast.AssignmentStatement); ok {
 		g.gen(&assStmt)
@@ -161,60 +95,6 @@ func (g *Generator) genAssStmt(node *ast.AssignmentStatement) string {
 
 	g.genAssignment(identifier)
 
-	return ""
-}
-
-func (g *Generator) genExpr(node *ast.Expression) string {
-	g.genExprOperand(&node.FirstOperand)
-	g.genExprOperand(&node.SecondOperand)
-
-	exprOp := string(node.Operator.(*token.Token).Lit)
-
-	switch exprOp {
-	case "+":
-		g.write("i32.add\n")
-		break
-	case "-":
-		g.write("i32.sub\n")
-		break
-	case "*":
-		g.write("i32.mul\n")
-		break
-	case "/":
-		g.write("i32.div_s\n")
-		break
-	default:
-		slog.Error("Failed to generate unknown operator")
-	}
-	return ""
-}
-
-func (g *Generator) genExprOperand(node *ast.Attribute) {
-	if operand, ok := (*node).(ast.Expression); ok {
-		g.genExpr(&operand)
-	} else {
-		g.genValue(node)
-	}
-}
-
-func (g *Generator) genValue(node *ast.Attribute) string {
-	if identifier, ok := g.isIdentifier(node); ok {
-		g.genVarUsage(&identifier)
-	}
-	if val, ok := (*node).(int64); ok {
-		g.genInt(val)
-	}
-	return ""
-}
-
-func (g *Generator) genVarUsage(node *Identifier) string {
-	identifier, ok := (*node).(Identifier)
-	if !ok {
-		slog.Error("Failed to generate identifier")
-	}
-
-	label := identifier.getOperation()
-	g.write("%s\n", label)
 	return ""
 }
